Refactor OAuth clients and refine auto refresh
