+++ foundry_sdk/_core/public_client_auth.py
@@ -13,7 +13,6 @@
 #  limitations under the License.
 
 
-import threading
 import time
 from typing import List
 from typing import Optional
@@ -51,18 +50,14 @@ class PublicClientAuth(OAuth):
     ) -> None:
         self._client_id = client_id
         self._redirect_url = redirect_url
-
-        self._token: Optional[OAuthToken] = None
-        self._should_refresh = should_refresh
-        self._stop_refresh_event = threading.Event()
-        self._scopes = scopes
         self._auth_request: Optional[AuthorizeRequest] = None
-        self._server_oauth_flow_provider = PublicClientOAuthFlowProvider(
+
+        server_oauth_flow_provider = PublicClientOAuthFlowProvider(
             client_id=client_id,
             redirect_url=redirect_url,
             scopes=scopes,
         )
-        super().__init__(hostname=hostname, should_refresh=should_refresh, config=config)
+        super().__init__(server_oauth_flow_provider=server_oauth_flow_provider, hostname=hostname, should_refresh=should_refresh, config=config)
 
     @property
     def scopes(self) -> List[str]:
@@ -73,17 +68,14 @@ class PublicClientAuth(OAuth):
             raise NotAuthenticated("Client has not been authenticated.")
         return self._token
 
-    def _refresh_token(self) -> None:
-        if not self._token:
-            raise RuntimeError("must have token to refresh")
-
-        if not self._token.refresh_token:
-            raise RuntimeError("no refresh token provided")
-
-        self._token = self._server_oauth_flow_provider.refresh_token(
-            self._get_client(),
-            refresh_token=self._token.refresh_token,
-        )
+    def _try_refresh_token(self) -> bool:
+        if self._token and self._token.refresh_token:
+            self._token = self._server_oauth_flow_provider.refresh_token(
+                self._get_client(),
+                refresh_token=self._token.refresh_token,
+            )
+            return True
+        return False
 
     @property
     def url(self) -> str:
@@ -94,23 +86,6 @@ class PublicClientAuth(OAuth):
             self._get_client()
         )
         return self._auth_request.url
-
-    def _start_auto_refresh(self) -> None:
-        def _auto_refresh_token() -> None:
-            while not self._stop_refresh_event.is_set():
-                if self._token and self._token.refresh_token:
-                    # Sleep for (expires_in - 60) seconds to refresh the token 1 minute before it expires
-                    time.sleep(self._token.expires_in - 60)
-                    self._token = self._server_oauth_flow_provider.refresh_token(
-                        self._get_client(),
-                        refresh_token=self._token.refresh_token,
-                    )
-                else:
-                    # Wait 10 seconds and check again if the token is set
-                    time.sleep(10)
-
-        refresh_thread = threading.Thread(target=_auto_refresh_token, daemon=True)
-        refresh_thread.start()
 
     def set_token(self, code: str, state: str) -> None:
         if not self._auth_request:
@@ -127,17 +102,3 @@ class PublicClientAuth(OAuth):
 
         if self._should_refresh:
             self._start_auto_refresh()
-
-    def sign_out(self) -> SignOutResponse:
-        if self._token:
-            self._server_oauth_flow_provider.revoke_token(
-                self._get_client(),
-                self._token.access_token,
-            )
-
-        self._token = None
-
-        # Signal the auto-refresh thread to stop
-        self._stop_refresh_event.set()
-
-        return SignOutResponse()
